requirements.txt
appdirs==1.4.4
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
cycler==0.10.0
decorator==4.4.2
Flask==2.0.1
Flask-Cors==3.0.10
future==0.18.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
meshio==4.4.3
moviepy==1.0.3
numpy==1.20.3
pathlib2==2.3.5
Pillow==8.2.0
proglog==0.1.9
pyparsing==2.4.7
python-dateutil==2.8.1
pyvista==0.30.1
requests==2.25.1
scipy==1.6.3
scooby==0.5.7
six==1.16.0
tqdm==4.61.0
transforms3d==0.3.1
urllib3==1.26.4
vtk==9.0.1
Werkzeug==2.0.1
